rag_pdf_agent/requirements.txt
google-generativeai>=0.3.0
pypdf>=3.17.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
tqdm>=4.66.0
torch>=2.0.0
